Refactor: minor fixes and doc file updation
