--- cloudKitchenBackend/src/main/java/com/cloudkitchenbackend/service/HomeService.java
@@ -1,7 +1,0 @@
-package com.cloudkitchenbackend.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class HomeService {
-}
