+++ cloudKitchenBackend/src/main/java/com/cloudkitchenbackend/service/MenuService.java
@@ -1,6 +1,5 @@
 package com.cloudkitchenbackend.service;
 
-import com.cloudkitchenbackend.controller.HomeController;
 import com.cloudkitchenbackend.exception.ItemAlreadyExistsException;
 import com.cloudkitchenbackend.exception.ItemNotFoundException;
 import com.cloudkitchenbackend.model.Item;
